Module volta_layout_macro::ast

Structs

Ast

Abstract syntax tree (AST) for the surface syntax of the layout! macro.

Directory

Represents a directory entry in the AST, which can recursively contain more entries.

FieldPrefix

AST for the common prefix of a single field in a layout! struct declaration, which is of the form:

LayoutStruct

Represents a single type LayoutStruct in the AST, which takes the form:

Enums

EntryKind

FieldContents

AST for the suffix of a field in a layout! struct declaration.

Type Definitions

Result